% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_gates.R
\name{calculate_gates}
\alias{calculate_gates}
\title{Calculate gates based on gate co-ordinates in csv file}
\usage{
calculate_gates(input_dir)
}
\arguments{
\item{input_dir}{directory which contains the file \code{gates_coordinates.csv}}
}
\value{
\code{list} containing two objects: gates for bacteria and gates for algae
}
\description{
Calculate gates based on gate co-ordinates in csv file
}
