% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dens.R
\name{dens}
\alias{dens}
\title{Get density using \code{area} for gating}
\usage{
dens(gates_coordinates, fsa, flow.data, min_FSC.A = 1e-10)
}
\arguments{
\item{gates_coordinates}{gates co-ordinates as read from the file \code{gates_coordinates.csv}}

\item{fsa}{as read from \code{flowcytometer_fsa_ungated.rds}}

\item{flow.data}{as read from \code{flowcytometer_ungated.csv}}

\item{min_FSC.A}{numeric. If \code{!NULL}, \code{FSA.A <= min_FSC.A} will be fitered out by using
a rectangular filter
\code{flowCore::rectangleGate(filterId="filter_out_0", "FSC-A" = c(min_FSC.A, +Inf))}}
}
\value{
gated \code{flow.data} density as saved in \code{flowcytometer_density.csv} and actual gates used
}
\description{
Get density using \code{area} for gating
}
