% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dens_H.R
\name{dens_H}
\alias{dens_H}
\title{Get density using \code{height} for gating}
\usage{
dens_H(gates_coordinates, fsa, flow.data, excl_FSCA_0 = TRUE)
}
\arguments{
\item{gates_coordinates}{gates co-ordinates as read from the file \code{gates_coordinates.csv}}

\item{fsa}{as read from \code{flowcytometer_fsa_ungated.rds}}

\item{flow.data}{as read from \code{flowcytometer_ungated.csv}}

\item{excl_FSCA_0}{boolean. If \code{TRUE}, \code{FSA.A <= 0} will be fitered out by using
a rectangular filter
\code{flowCore::rectangleGate(filterId="filter_out_0", "FSC-A" = c(0.00000000001, +Inf))}}
}
\value{
gated \code{flow.data} density as saved in \code{flowcytometer_density.csv}
}
\description{
Get density using \code{height} for gating
}
