% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractor_flowcytometer_density.R
\name{extractor_flowcytometer_density}
\alias{extractor_flowcytometer_density}
\title{Extractor flowcytometer data}
\usage{
extractor_flowcytometer_density(
  input = NA,
  output = ".",
  min_FSC.A = NULL,
  use_H = FALSE,
  gates_coordinates = NULL,
  fsa = NULL,
  flow.data = NULL,
  dens_back = FALSE
)
}
\arguments{
\item{input}{directory from which to read the data}

\item{output}{directory to which to write the data}
}
\value{
invisibly \code{TRUE} when completed successful
}
\description{
This function is extracting data to be added to the database
(and therefore make accessible for further analysis and forecasting)
from \code{.fcs} files.
}
