% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractor_traits.R
\name{extractor_traits}
\alias{extractor_traits}
\title{Extractor flowcytometer data}
\usage{
extractor_traits(
  input = NULL,
  particles = c("bacteria", "LNA", "MNA", "HNA", "algae"),
  metadata_flowcytometer,
  min_FSC.A = NULL,
  use_H = FALSE,
  timestamp = yaml::read_yaml(file.path(input, "sample_metadata.yml"))$timestamp,
  fsa_p1 = NULL,
  fsa_p2 = NULL,
  gates_coordinates = NULL,
  wellid_keyword = "$WELLID"
)
}
\arguments{
\item{input}{directory from which to read the data}

\item{metadata_flowcytometer}{the content of the file
\code{metadata_flowcytometer.csv} which will be linked into the traits}

\item{min_FSC.A}{numeric. If \code{!NULL}, \code{FSA.A <= min_FSC.A} will be fitered out by using
a rectangular filter
\code{flowCore::rectangleGate(filterId="filter_out_0", "FSC-A" = c(min_FSC.A, +Inf))}}

\item{use_H}{if \code{TRUE}, gating will be done using \code{height},
otherwie \code{area}}

\item{timestamp}{timestamp. Default: read from \code{sample_metadata.yml}}

\item{fsa_p1}{if \code{NULL}, \code{fsa_p1} will be read in, otherwise the
\code{fsa}}

\item{fsa_p2}{if \code{NULL}, \code{fsa_p2} will be read in, otherwise the
\code{fsa}}

\item{gates_coordinates}{if \code{NULL}, \code{gates_coordinates} will be
read in, otherwise the \code{gates_coordinates}}

\item{wellid_keyword}{the kwyword which is used to identify the well ID.
Usually "$WELLID" (default), but for the EAWAG Flowcytometer it is "$SMNO".}
}
\value{
invisibly \code{TRUE} when completed successful
}
\description{
This function is extracting data to be added to the database
(and therefore make accessible for further analysis and forecasting)
from \code{.fcs} files.
}
