% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating.R
\name{gating}
\alias{gating}
\title{Do the gating}
\usage{
gating(gates_coordinates, fsa, flow.data)
}
\arguments{
\item{gates_coordinates}{gates co-ordinates as read from the file \code{gates_coordinates.csv}}

\item{fsa}{as read from \code{flowcytometer_fsa_ungated.rds}}

\item{flow.data}{as read from \code{flowcytometer_ungated.csv}}
}
\value{
gated \code{flow.data} density as saved in \code{flowcytometer_density.csv} and actual gates used
}
\description{
Do the gating
}
