% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processor_flowcytometer.R
\name{pre_processor_flowcytometer}
\alias{pre_processor_flowcytometer}
\title{Preprocessor flowcytometer data}
\usage{
pre_processor_flowcytometer(input, output)
}
\arguments{
\item{input}{directory from which to read the data}

\item{output}{directory to which to write the data}
}
\value{
invisibly \code{TRUE} when completed successful
}
\description{
Just copy all files
}
