% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processor_flowcytometer_c6.R
\name{pre_processor_flowcytometer_c6}
\alias{pre_processor_flowcytometer_c6}
\title{Preprocessor flowcytometer data FOR c6 FILES!!!}
\usage{
pre_processor_flowcytometer_c6(input, output)
}
\arguments{
\item{input}{directory from which to read the data}

\item{output}{directory to which to write the data}
}
\value{
invisibly \code{TRUE} when completed successful
}
\description{
Convert all \code{.c6} files in \code{flowcytometrie} folder to \code{.fcs} files in output folder
}
