% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{register}
\alias{register}
\title{Register the processing of flowcytometer data in the LEEF.Data package}
\usage{
register()
}
\value{
invisibly \code{TRUE} when completed successful
}
\description{
Register the processing of flowcytometer data in the LEEF.Data package
}
