% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_gates.R
\name{calculate_gates}
\alias{calculate_gates}
\title{Calculate gates based on gate co-ordinates in csv file using \code{area}}
\usage{
calculate_gates(input_dir = NULL, gates_coordinates = NULL)
}
\arguments{
\item{input_dir}{directory which contains the file \code{gates_coordinates.csv}}

\item{gates_coordinates}{contains gate co-ordinates. If \code{NULL} *default)
the gates cootdinates are read in from \code{input_dir}}
}
\value{
\code{list} containing two objects: gates for bacteria and gates for algae
}
\description{
Calculate gates based on gate co-ordinates in csv file using \code{area}
}
