% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_traits.R
\name{extract_traits}
\alias{extract_traits}
\title{Extractor flowcytometer data}
\usage{
extract_traits(
  input,
  particles = c("bacteria", "LNA", "MNA", "HNA", "algae"),
  metadata_flowcytometer
)
}
\arguments{
\item{input}{directory from which to read the data}

\item{particles}{\code{character} vector containing the groups to extract.
Supported are \code{"bacteria"}, \code{"LNA"}, \code{"MNA"}, \code{"HNA"},
\code{"algae"}.}

\item{metadata_flowcytometer}{the content of the file \code{metadata_flowcytometer.csv} which will be linked into the traits}
}
\value{
invisibly \code{TRUE} when completed successful
}
\description{
This function is extracting data to be added to the database
(and therefore make accessible for further analysis and forecasting)
from \code{.fcs} files.
}
