% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractor_flowcytometer_density.R
\name{extractor_flowcytometer_density}
\alias{extractor_flowcytometer_density}
\title{Extractor flowcytometer density}
\usage{
extractor_flowcytometer_density(
  input = NULL,
  output = ".",
  min_FSC.A = NULL,
  use_H = FALSE,
  gates_coordinates = NULL,
  fsa = NULL,
  flow.data = NULL,
  dens_back = FALSE
)
}
\arguments{
\item{input}{directory from which to read the data}

\item{output}{directory to which to write the data}

\item{min_FSC.A}{numeric. If \code{!NULL}, \code{FSA.A <= min_FSC.A} will be fitered out by using
a rectangular filter
\code{flowCore::rectangleGate(filterId="filter_out_0", "FSC-A" = c(min_FSC.A, +Inf))}}

\item{use_H}{if \code{TRUE}, gating will be done using \code{height}, otherwie \code{area}}

\item{gates_coordinates}{if \code{NULL}, \code{gates_coordinates} will be read in, otherwise the \code{gates_coordinates}}

\item{fsa}{if \code{NULL}, \code{fsa} will be read in, otherwise the \code{fsa}}

\item{flow.data}{if \code{NULL}, \code{flow.data} will be read in, otherwise the \code{flow.data}}

\item{dens_back}{if \code{FALSE}, density will be saved in "flowcytometer_density.csv", otherwise it will be returned}
}
\value{
invisibly \code{TRUE} when completed successful
}
\description{
This function is extracting data to be added to the database
(and therefore make accessible for further analysis and forecasting)
from \code{.fcs} files.
}
