% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractor_flowcytometer_density.R
\name{extractor_flowcytometer_density}
\alias{extractor_flowcytometer_density}
\title{Extractor flowcytometer density}
\usage{
extractor_flowcytometer_density(
  input,
  output,
  excl_FSCA_0 = FALSE,
  use_H = FALSE
)
}
\arguments{
\item{input}{directory from which to read the data}

\item{output}{directory to which to write the data}

\item{excl_FSCA_0}{boolean. If \code{TRUE}, \code{FSA.A <= 0} will be fitered out by using
a rectangular filter
\code{flowCore::rectangleGate(filterId="filter_out_0", "FSC-A" = c(0.00000000001, +Inf))}}

\item{use_H}{if \code{TRUE}, gating will be done using \code{height}, otherwie \code{area}}
}
\value{
invisibly \code{TRUE} when completed successful
}
\description{
This function is extracting data to be added to the database
(and therefore make accessible for further analysis and forecasting)
from \code{.fcs} files.
}
