% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractor_flowcytometer_preparation.R
\name{extractor_flowcytometer_preparation}
\alias{extractor_flowcytometer_preparation}
\title{Extractor flowcytometer preparation data}
\usage{
extractor_flowcytometer_preparation(input, output, raw = FALSE)
}
\arguments{
\item{input}{directory from which to read the data}

\item{output}{directory to which to write the data}
}
\value{
invisibly \code{TRUE} when completed successful
}
\description{
This function is extracting data to be added to the database
(and therefore make accessible for further analysis and forecasting)
from \code{.fcs} files.
}
